making nicer code
